+++ npm-packages/react-api/src/main.ts
@@ -2,16 +2,18 @@ import useAgentTools from "./hooks/agent/use-agent-tools";
 import useAgents from "./hooks/agent/use-agents";
 import useFetch from "./hooks/editor/use-fetch";
 import useFileView from "./hooks/editor/use-file-view";
+import useLoading from "./hooks/editor/use-loading";
 import useNotification from "./hooks/editor/use-notification";
 import useTheme from "./hooks/editor/use-theme";
 import useToolbar from "./hooks/editor/use-toolbar";
 import useExtCommand from "./hooks/extension/use-ext-command";
 
-import useDiffusion from "./hooks/modality/use-diffusion";
+import useImageGen from "./hooks/modality/use-image-gen";
 import useLLM from "./hooks/modality/use-llm";
 import useOCR from "./hooks/modality/use-ocr";
 import useSTT from "./hooks/modality/use-stt";
 import useTTS from "./hooks/modality/use-tts";
+import useVideoGen from "./hooks/modality/use-video-gen";
 import useTerminal from "./hooks/terminal/use-terminal";
 
 export {
@@ -22,11 +24,13 @@ export {
   useNotification,
   useTheme,
   useToolbar,
-  useDiffusion,
+  useImageGen,
+  useVideoGen,
   useLLM,
   useOCR,
   useSTT,
   useTTS,
   useExtCommand,
   useTerminal,
+  useLoading,
 };