+++ web/components/providers/imc-provider.tsx
@@ -2,6 +2,7 @@
 
 import { IMCContextType, PlatformEnum } from "@/lib/types";
 import {
+  ImageModelConfig,
   IMCMessage,
   IMCMessageTypeEnum,
   LLMConfig,
@@ -10,16 +11,25 @@ import {
   STTConfig,
   TTSConfig,
 } from "@pulse-editor/shared-utils";
-import { createContext, useContext, useEffect, useState } from "react";
+import { createContext, useContext, useEffect, useRef, useState } from "react";
 import { EditorContext } from "./editor-context-provider";
 import { getPlatform } from "@/lib/platform-api/platform-checker";
 import { usePlatformApi } from "@/lib/hooks/use-platform-api";
 import { getAPIKey } from "@/lib/settings/settings";
 import { runAgentMethod } from "@/lib/agent/agent-runner";
-import { getModelLLM } from "@/lib/llm/llm";
-import { getModelTTS } from "@/lib/tts/tts";
-import { getModelSTT } from "@/lib/stt/stt";
-import { recognizeText } from "@/lib/image-processing/ocr";
+import { getLLMModel } from "@/lib/modalities/llm/llm";
+import { getTTSModel } from "@/lib/modalities/tts/tts";
+import { getSTTModel } from "@/lib/modalities/stt/stt";
+import { recognizeText } from "@/lib/modalities/ocr/ocr";
+import {
+  getDefaultImageModelConfig,
+  getDefaultLLMConfig,
+  getDefaultSTTConfig,
+  getDefaultTTSConfig,
+  getDefaultVideoModelConfig,
+} from "@/lib/modalities/utils";
+import { getImageGenModel } from "@/lib/modalities/image-gen/image-gen";
+import { getVideoGenModel } from "@/lib/modalities/video-gen/video-gen";
 
 export const IMCContext = createContext<IMCContextType | undefined>(undefined);
 
@@ -36,6 +46,14 @@ export default function InterModuleCommunicationProvider({
   useEffect(() => {
     // @ts-expect-error set window viewId
     window.viewId = "Pulse Editor Main";
+
+    return () => {
+      // Cleanup the polyIMC instance when the component unmounts
+      if (polyIMC) {
+        polyIMC.close();
+        setPolyIMC(undefined);
+      }
+    };
   }, []);
 
   useEffect(() => {
@@ -178,7 +196,8 @@ export default function InterModuleCommunicationProvider({
             sttConfig,
           }: { audio: ArrayBuffer; sttConfig?: STTConfig } = message.payload;
 
-          const config = sttConfig ? sttConfig : undefined; // TODO: use editor level default config -- getDefaultLLMConfig();
+          const config = sttConfig ?? getDefaultSTTConfig(editorContext);
+
           if (!config) {
             throw new Error("No STT config found for this agent method.");
           }
@@ -189,7 +208,7 @@ export default function InterModuleCommunicationProvider({
             throw new Error(`No API key found for provider ${provider}.`);
           }
 
-          const stt = getModelSTT(apiKey, provider, config.modelName);
+          const stt = getSTTModel(apiKey, provider, config.modelName);
           if (!stt) {
             throw new Error("STT not found.");
           }
@@ -219,7 +238,7 @@ export default function InterModuleCommunicationProvider({
             llmConfig?: LLMConfig;
           } = message.payload;
 
-          const config = llmConfig ? llmConfig : undefined; // TODO: use editor level default config -- getDefaultLLMConfig();
+          const config = llmConfig ?? getDefaultLLMConfig(editorContext);
 
           if (!config) {
             throw new Error("No LLM config found for this agent method.");
@@ -232,7 +251,7 @@ export default function InterModuleCommunicationProvider({
             throw new Error(`No API key found for provider ${provider}.`);
           }
 
-          const llm = getModelLLM(
+          const llm = getLLMModel(
             apiKey,
             provider,
             config.modelName,
@@ -261,7 +280,7 @@ export default function InterModuleCommunicationProvider({
           const { text, ttsConfig }: { text: string; ttsConfig?: TTSConfig } =
             message.payload;
 
-          const config = ttsConfig ? ttsConfig : undefined; // TODO: use editor level default config -- getDefaultLLMConfig();
+          const config = ttsConfig ?? getDefaultTTSConfig(editorContext);
 
           if (!config) {
             throw new Error("No TTS config found for this agent method.");
@@ -273,7 +292,7 @@ export default function InterModuleCommunicationProvider({
             throw new Error(`No API key found for provider ${provider}.`);
           }
 
-          const tts = getModelTTS(
+          const tts = getTTSModel(
             apiKey,
             provider,
             config.modelName,
@@ -291,7 +310,7 @@ export default function InterModuleCommunicationProvider({
         },
       ],
       [
-        IMCMessageTypeEnum.UseDiffusion,
+        IMCMessageTypeEnum.UseImageGen,
         async (
           senderWindow: Window,
           message: IMCMessage,
@@ -299,11 +318,87 @@ export default function InterModuleCommunicationProvider({
         ) => {
           // Handle the use diffusion message
           console.log(
-            "Received use diffusion message from extension:",
+            `Received ${IMCMessageTypeEnum.UseImageGen.toString()} message from extension:`,
             message,
           );
 
-          throw new Error("Not implemented");
+          const {
+            textPrompt,
+            imagePrompt,
+            imageModelConfig,
+          }: {
+            textPrompt?: string;
+            imagePrompt?: string | ArrayBuffer;
+            imageModelConfig?: ImageModelConfig;
+          } = message.payload;
+
+          const config =
+            imageModelConfig ?? getDefaultImageModelConfig(editorContext);
+
+          if (!config) {
+            throw new Error(
+              "No image model config found for this agent method.",
+            );
+          }
+
+          const provider = config.provider;
+          const apiKey = getAPIKey(editorContext, provider);
+          if (!apiKey) {
+            throw new Error(`No API key found for provider ${provider}.`);
+          }
+
+          const model = getImageGenModel(apiKey, provider, config.modelName);
+
+          const res = await model.generate(textPrompt, imagePrompt);
+
+          return res;
+        },
+      ],
+      [
+        IMCMessageTypeEnum.UseVideoGen,
+        async (
+          senderWindow: Window,
+          message: IMCMessage,
+          abortSignal?: AbortSignal,
+        ) => {
+          // Handle the use video generation message
+          console.log(
+            `Received ${IMCMessageTypeEnum.UseVideoGen.toString()} message from extension:`,
+            message,
+          );
+
+          const {
+            duration,
+            textPrompt,
+            imagePrompt,
+            videoModelConfig,
+          }: {
+            duration: number;
+            textPrompt?: string;
+            imagePrompt?: string | ArrayBuffer;
+            videoModelConfig?: ImageModelConfig;
+          } = message.payload;
+
+          const config =
+            videoModelConfig ?? getDefaultVideoModelConfig(editorContext);
+
+          if (!config) {
+            throw new Error(
+              "No video model config found for this agent method.",
+            );
+          }
+
+          const provider = config.provider;
+          const apiKey = getAPIKey(editorContext, provider);
+          if (!apiKey) {
+            throw new Error(`No API key found for provider ${provider}.`);
+          }
+
+          const model = getVideoGenModel(apiKey, provider, config.modelName);
+
+          const res = await model.generate(duration, textPrompt, imagePrompt);
+
+          return res;
         },
       ],
       [