+++ web/components/views/loaders/view-loader.tsx
@@ -115,6 +115,7 @@ export default function ViewLoader({
 
     if (currentViewId) {
       // Listen for an incoming extension connection
+      console.log("Listening for extension connection...");
       listenForExtensionConnection();
 
       setIsLookingForExtension(true);
@@ -174,13 +175,18 @@ export default function ViewLoader({
 
     // Add loaded handler
     newMap.set(
-      IMCMessageTypeEnum.Loaded,
+      IMCMessageTypeEnum.UseLoading,
       async (
         senderWindow: Window,
         message: IMCMessage,
         abortSignal?: AbortSignal,
       ) => {
-        setIsLoadingExtension((prev) => false);
+        const {
+          isLoading,
+        }: {
+          isLoading: boolean;
+        } = message.payload;
+        setIsLoadingExtension((prev) => isLoading);
       },
     );