+++ npm-packages/shared-utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @pulse-editor/shared-utils
 
+## 0.1.1-alpha.18
+
+### Patch Changes
+
+- Add image and video gen hooks
+
+## 0.1.1-alpha.17
+
+### Patch Changes
+
+- Add useLoading, image gen, video gen hooks
+
 ## 0.1.1-alpha.16
 
 ### Patch Changes